# ENH: linalg: speedup _sqrtm_triu by moving tight loop to Cython