--- a/scipy/linalg/setup.py
+++ b/scipy/linalg/setup.py
@@ -105,6 +105,11 @@ def configuration(parent_package='', top_path=None):
                          sources=[('_solve_toeplitz.c')],
                          include_dirs=[get_numpy_include_dirs()])
 
+    # _matfuncs_sqrtm_triu:
+    config.add_extension('_matfuncs_sqrtm_triu',
+                         sources=[('_matfuncs_sqrtm_triu.c')],
+                         include_dirs=[get_numpy_include_dirs()])
+
     config.add_data_dir('tests')
 
     # Cython BLAS/LAPACK